Add support for searching by usage rights
A common search option, conveniently exposed by the google image
search image, is to filter by copyright usage.

As part of the advanced options, which is where the aspect ratio
and image type are bundled is the usage rights option, below are
the values for each option:

  sur:fmc (labeled for reuse with modification)
  sur:fc (labeled for reuse)
  sur:fm (labeled for noncommercial reuse with modification)
  sur:f (labeled for noncommercial reuse)

This PR adds a new option, --usage-rights <usage-rights> to the
CLI, with a few preset values for users to choose based on the
options mentioned above.